Share the metro config solving symbolic links
