+++ integration-tests/environments/react-native/package.json
@@ -40,6 +40,7 @@
   "devDependencies": {
     "@babel/core": "^7.12.9",
     "@babel/runtime": "^7.12.5",
+    "@realm/metro-config": "*",
     "concurrently": "^6.0.2",
     "metro-react-native-babel-preset": "^0.66.2",
     "mocha-github-actions-reporter": "^0.2.3",
