+++ packages/metro-config/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+  "env": {
+    "node": true
+  }
+}
