+++ packages/metro-config/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@realm/metro-config",
+  "version": "0.1.0",
+  "description": "Metro config shared between our sub-packages, to support symbolic links",
+  "main": "index.js",
+  "dependencies": {
+    "metro-config": "^0.66.2"
+  }
+}
